Split deploy and provide MUNI LP into 2 scripts
